fixed logical conflicts
