+++ src/main/java/org/umssdiplo/automationv01/core/managepage/editItemType/EditItemType.java
@@ -27,12 +27,14 @@ public class EditItemType extends BasePage {
 
     public ListItemType clickSaveItemTypeBtn() {
         CommonEvents.customWait(2500);
+        CommonEvents.mouseOver(saveItemTypeBtn);
         CommonEvents.clickButton(saveItemTypeBtn);
         return new ListItemType();
     }
 
     public ListItemType clickCancelEditItemTypeBtn() {
         CommonEvents.customWait(2500);
+        CommonEvents.mouseOver(cancelEditItemTypeBtn);
         CommonEvents.clickButton(cancelEditItemTypeBtn);
         return new ListItemType();
     }
