+++ src/main/java/org/umssdiplo/automationv01/core/managepage/navigationbar/NavigationBar.java
@@ -50,7 +50,7 @@ public class NavigationBar extends BasePage {
 
     public ListDepartmentPage clickOnSubmenuButton() {
         CommonEvents.isVisible(activeSubmenu);
-        CommonEvents.clickButton(activeSubmenu);
+        CommonEvents.jsClickElement(activeSubmenu);
         return new ListDepartmentPage();
     }
