+++ src/main/java/org/umssdiplo/automationv01/core/managepage/organizationalstructuremenu/OrganizationalStructureMenu.java
@@ -40,7 +40,7 @@ public class OrganizationalStructureMenu extends BasePage {
 
     public RegisterContract clickContracts() {
         CommonEvents.customWait(2500);
-        CommonEvents.clickButton(contractButton);
+        CommonEvents.jsClickElement(contractButton);
         return new RegisterContract();
     }
 }
